move debounce to vendor/debounce es6 module
